Add dedicated util.ts file, move first two methods
