+++ packages/client/src/rpc/modules/engine/engine.ts
@@ -4,7 +4,6 @@ import { BlobEIP4844Transaction } from '@ethereumjs/tx'
 import {
   BIGINT_0,
   BIGINT_1,
-  bigIntToHex,
   bytesToHex,
   bytesToUnprefixedHex,
   equalsBytes,
@@ -28,13 +27,8 @@ import { callWithStackTrace } from '../../helpers'
 import { CLConnectionManager, middleware as cmMiddleware } from '../../util/CLConnectionManager'
 import { middleware, validators } from '../../validation'
 
-import {
-  type BlobsBundleV1,
-  type ChainCache,
-  EngineError,
-  type PayloadStatusV1,
-  Status,
-} from './types'
+import { type ChainCache, EngineError, type PayloadStatusV1, Status } from './types'
+import { blockToExecutionPayload, pruneCachedBlocks } from './util'
 import {
   executionPayloadV1FieldValidators,
   executionPayloadV2FieldValidators,
@@ -49,7 +43,6 @@ import type { Chain } from '../../../blockchain'
 import type { EthereumClient } from '../../../client'
 import type { Config } from '../../../config'
 import type { VMExecution } from '../../../execution'
-import type { BlobsBundle } from '../../../miner'
 import type { FullEthereumService, Skeleton } from '../../../service'
 import type {
   Bytes32,
@@ -71,87 +64,6 @@ import type { Common } from '@ethereumjs/common'
 import type { VM } from '@ethereumjs/vm'
 
 const zeroBlockHash = zeros(32)
-
-/**
- * Formats a block to {@link ExecutionPayloadV1}.
- */
-export const blockToExecutionPayload = (block: Block, value: bigint, bundle?: BlobsBundle) => {
-  const blockJson = block.toJSON()
-  const header = blockJson.header!
-  const transactions = block.transactions.map((tx) => bytesToHex(tx.serialize())) ?? []
-  const withdrawalsArr = blockJson.withdrawals ? { withdrawals: blockJson.withdrawals } : {}
-  const blobsBundle: BlobsBundleV1 | undefined = bundle
-    ? {
-        commitments: bundle.commitments.map(bytesToHex),
-        blobs: bundle.blobs.map(bytesToHex),
-        proofs: bundle.proofs.map(bytesToHex),
-      }
-    : undefined
-
-  const executionPayload: ExecutionPayload = {
-    blockNumber: header.number!,
-    parentHash: header.parentHash!,
-    feeRecipient: header.coinbase!,
-    stateRoot: header.stateRoot!,
-    receiptsRoot: header.receiptTrie!,
-    logsBloom: header.logsBloom!,
-    gasLimit: header.gasLimit!,
-    gasUsed: header.gasUsed!,
-    timestamp: header.timestamp!,
-    extraData: header.extraData!,
-    baseFeePerGas: header.baseFeePerGas!,
-    blobGasUsed: header.blobGasUsed,
-    excessBlobGas: header.excessBlobGas,
-    blockHash: bytesToHex(block.hash()),
-    prevRandao: header.mixHash!,
-    transactions,
-    ...withdrawalsArr,
-  }
-
-  // ethereumjs doesnot provide any transaction censoring detection (yet) to suggest
-  // overriding builder/mev-boost blocks
-  const shouldOverrideBuilder = false
-  return { executionPayload, blockValue: bigIntToHex(value), blobsBundle, shouldOverrideBuilder }
-}
-
-const pruneCachedBlocks = (chain: Chain, chainCache: ChainCache) => {
-  const { remoteBlocks, executedBlocks, invalidBlocks } = chainCache
-  const finalized = chain.blocks.finalized
-  if (finalized !== null) {
-    // prune remoteBlocks
-    const pruneRemoteBlocksTill = finalized.header.number
-    for (const blockHash of remoteBlocks.keys()) {
-      const block = remoteBlocks.get(blockHash)
-      if (block !== undefined && block.header.number <= pruneRemoteBlocksTill) {
-        remoteBlocks.delete(blockHash)
-      }
-    }
-
-    // prune executedBlocks
-    const vm = chain.blocks.vm
-    if (vm !== null) {
-      const pruneExecutedBlocksTill =
-        vm.header.number < finalized.header.number ? vm.header.number : finalized.header.number
-      for (const blockHash of executedBlocks.keys()) {
-        const block = executedBlocks.get(blockHash)
-        if (block !== undefined && block.header.number < pruneExecutedBlocksTill) {
-          executedBlocks.delete(blockHash)
-        }
-      }
-    }
-
-    // prune invalidBlocks with some max length
-    const pruneInvalidLength = invalidBlocks.size - chain.config.maxInvalidBlocksErrorCache
-    let pruned = 0
-    for (const blockHash of invalidBlocks.keys()) {
-      if (pruned >= pruneInvalidLength) {
-        break
-      }
-      invalidBlocks.delete(blockHash)
-      pruned++
-    }
-  }
-}
 
 /**
  * Recursively finds parent blocks starting from the parentHash.
