+++ packages/client/src/rpc/modules/engine/index.ts
@@ -1,3 +1,4 @@
 export * from './engine'
 export * from './types'
+export * from './util'
 export * from './validators'
